Refactoring of BeanTestExtension: extracted the Interceptor/EJB functionality in seperate CDI Extensions
This commit provides a spreading of the former BeanTestExtension in seperate CDI Extensions. Namely: BaseExtension, EJBExtension and InterceptorExtension.
Further this commit contains various cleanups.

Details:

* Introduced Base class for CDI Extensions
* Created distinct CDI Extensions and registered in the javax.enterprise.inject.spi.Extension file
* Removed redundant type parameter from methods
* Pulled up various common methods in BaseExtension
* Updated Javadocs
