+++ src/test/java/info/novatec/beantest/demo/ejb/InvalidInjectionPointConfigurationEJB.java
@@ -16,15 +16,13 @@
 
 package info.novatec.beantest.demo.ejb;
 
-import info.novatec.beantest.extension.BeanTestExtension;
-
 import javax.ejb.Stateless;
 import javax.enterprise.inject.Vetoed;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
 /**
- * This EJB holds an invalid dependency configuration in order to test the deployment functionality of {@link BeanTestExtension}.
+ * This EJB holds an invalid dependency configuration in order to test the deployment functionality of {@link info.novatec.beantest.extension.BaseExtension}.
  * <p>
  * This bean is currently vetoed due to the fact that this bean definition causes a deployment exception during the bean container bootstrapping phase.
  * As soon there is a solution to enable alternative beans per test case this bean altogether with its corresponding test case will be reactivated.
